feat: added network metric; latency, bandwidth usage, etc
