+++ integration/src/mastra/agents/metricsAgent.ts
@@ -18,6 +18,7 @@ IMPORTANT: You must ONLY respond with ONE of these exact keywords - nothing more
 - "memoryStats" - When user asks about memory usage, RAM, memory status, or anything memory-related
 - "diskMetrics" - When user asks about disk usage, storage, disk I/O, read/write speeds, disk operations, or any disk-related metrics
 - "processMetrics" - When user asks about running processes, top processes, CPU-intensive processes, process count, zombie processes, or any process-related metrics
+- "networkMetrics" - When user asks about network usage, bandwidth, latency, packet loss, network interfaces, or any network-related metrics
 - "getAllMetrics" - When user asks for all metrics, overall server status, or a complete overview
 - "setup-monitoring" - When user asks about setting up, installing, or configuring monitoring
 - "conversation" - For ANY message that doesn't directly request metrics data:
@@ -38,6 +39,9 @@ Examples:
 - "what processes are running?" → "processMetrics"
 - "show me top CPU processes" → "processMetrics"
 - "any zombie processes?" → "processMetrics"
+- "check network bandwidth" → "networkMetrics"
+- "show me latency" → "networkMetrics"
+- "what's the packet loss?" → "networkMetrics"
 - "hello there" → "conversation"
 - "thank you for the info" → "conversation"
 - "what does load average mean?" → "conversation"
