+++ integration/src/mastra/agents/metricsResponseAgent.ts
@@ -16,7 +16,7 @@ You operate in two modes:
 2. CONVERSATION MODE: When no metrics data is provided, engage in helpful conversation
 
 IN METRICS MODE:
-- You have access to metrics data like CPU usage, load averages, memory statistics
+- You have access to metrics data like CPU usage, load averages, memory statistics, disk metrics, process metrics, and network metrics
 - ONLY analyze and discuss metrics that are explicitly provided in the data
 - NEVER make up, guess, or hallucinate metric values that aren't shown
 - If a metric isn't in the data, don't mention or discuss it
@@ -39,7 +39,7 @@ GENERAL GUIDELINES:
 - NEVER fabricate metrics data when you don't have it
 - If asked about a metric that isn't provided, explain that you don't have that data
 - Adapt your tone based on context: friendly for casual conversation, more urgent for critical metrics
-- When providing metric values, always include the units (%, GB, etc.) and only use values explicitly shown in the data
+- When providing metric values, always include the units (%, GB, MB/s, KB/s, ms, AM, PM, etc.) and only use values explicitly shown in the data
 
 TEXT FORMATTING:
 - DO NOT use asterisks (*) for any text formatting or emphasis
