+++ package/src/metrics/collector.ts
@@ -1,6 +1,10 @@
 import os from "os";
 import si from "systeminformation";
 import { logger } from "../utils/logger.js";
+import { getProcessMetrics } from "./processes.js";
+import { getDiskMetrics } from "./disk.js";
+import { getCpuMetrics, getCpuUsagePerCoreMetrics } from "./cpu.js";
+import { getNetworkMetrics } from "./network.js";
 
 // Define the metrics data structure
 export interface IMetricsData {
@@ -61,178 +65,24 @@ export interface IMetricsData {
       command: string; // command used to start the process
     }>;
   };
-}
-
-// get cpu and memory metrics
-async function getCpuMetrics(): Promise<Partial<IMetricsData>> {
-  try {
-    const [currentLoad, cpuInfo] = await Promise.all([
-      si.currentLoad(),
-      si.cpu(),
-    ]);
-
-    const load = await si.currentLoad();
-
-    // CPU Load average
-    const loadAverages = os.loadavg();
-    const numCores = cpuInfo.cores;
-
-    const normalizedLoad = {
-      "1min": (loadAverages[0] / numCores) * 100,
-      "5mins": (loadAverages[1] / numCores) * 100,
-      "15mins": (loadAverages[2] / numCores) * 100,
+  networkMetrics?: {
+    timestamp: number;
+    bandwidthUsage: {
+      received: number; // bytes per second
+      sent: number; // bytes per second
+      lastMeasurement: number;
     };
-
-    // CPU memory
-    const mem = await si.mem();
-    const usedGB = mem.used / 1024 ** 3; // Convert to GB
-    const totalGB = mem.total / 1024 ** 3;
-    const percentUsed = (mem.used / mem.total) * 100;
-
-    const memData = {
-      used: usedGB,
-      total: totalGB,
-      percentage: percentUsed,
+    latency: number; // milliseconds
+    packetLoss: number; // percentage
+    interfaceStats: {
+      [interfaceName: string]: {
+        rxBytes: number;
+        txBytes: number;
+        errors: number;
+      };
     };
-
-    return {
-      cpu: {
-        usage: currentLoad.currentLoad,
-        cores: cpuInfo.cores,
-        load_avg: [load.avgLoad],
-      },
-      cpuLoadAvgs: normalizedLoad,
-      memory: memData,
-    };
-  } catch (error) {
-    logger.error(`Failed to get CPU metrics: ${(error as Error).message}`);
-    throw error;
-  }
-}
-
-// get the cpu usage per core
-async function getCpuUsagePerCoreMetrics(): Promise<Partial<IMetricsData>> {
-  try {
-    const cl = await si.currentLoad();
-    const cpus = cl.cpus;
-    const coresLoad = cpus.map((cpu) => cpu.load);
-    return { cpuUsagePerCore: coresLoad };
-  } catch (error) {
-    logger.error(
-      `Failed to get CPU usage per core: ${(error as Error).message}`
-    );
-    throw error;
-  }
-}
-
-// get disk metrics
-async function getDiskMetrics(): Promise<Partial<IMetricsData>> {
-  try {
-    // Get file system information
-    const fsSize = await si.fsSize();
-
-    // Get disk I/O information
-    const diskIO = await si.disksIO();
-
-    // Format the disk information
-    const filesystems = fsSize.map((fs) => ({
-      fs: fs.fs,
-      type: fs.type,
-      size: fs.size,
-      used: fs.used,
-      available: fs.available,
-      use: fs.use,
-      mount: fs.mount,
-    }));
-
-    // Format the I/O information
-    // Note: diskIO returns data for all disks combined
-    const io = {
-      rIO: diskIO.rIO || 0,
-      wIO: diskIO.wIO || 0,
-      tIO: diskIO.tIO || 0,
-      rWaitTime: diskIO.rWaitTime || 0,
-      wWaitTime: diskIO.wWaitTime || 0,
-      tWaitTime: diskIO.tWaitTime || 0,
-      rPerSec: diskIO.rIO_sec || 0,
-      wPerSec: diskIO.wIO_sec || 0,
-    };
-
-    return {
-      disk: {
-        filesystems,
-        io,
-      },
-    };
-  } catch (error) {
-    logger.error(`Failed to get disk metrics: ${(error as Error).message}`);
-    throw error;
-  }
-}
-
-// get process metrics
-async function getProcessMetrics(): Promise<Partial<IMetricsData>> {
-  try {
-    // Get process information
-    const processes = await si.processes();
-
-    // Get the top 10 processes by CPU usage
-    const topProcessesByCpu = [...processes.list]
-      .sort((a, b) => b.cpu - a.cpu)
-      .slice(0, 10)
-      .map((proc) => ({
-        pid: proc.pid,
-        name: proc.name,
-        cpu: proc.cpu,
-        mem: proc.mem,
-        priority: proc.priority,
-        memVsz: proc.memVsz,
-        memRss: proc.memRss,
-        state: proc.state,
-        user: proc.user,
-        command: proc.command,
-      }));
-
-    // Process state counts
-    const stateCount = {
-      running: 0,
-      sleeping: 0,
-      blocked: 0,
-      zombie: 0,
-      unknown: 0,
-    };
-
-    // Count process states
-    processes.list.forEach((proc) => {
-      const state = proc.state.toLowerCase();
-      if (state.includes("running")) {
-        stateCount.running++;
-      } else if (state.includes("sleep")) {
-        stateCount.sleeping++;
-      } else if (state.includes("blocked") || state.includes("wait")) {
-        stateCount.blocked++;
-      } else if (state.includes("zombie")) {
-        stateCount.zombie++;
-      } else {
-        stateCount.unknown++;
-      }
-    });
-
-    return {
-      processes: {
-        all: processes.all,
-        running: stateCount.running,
-        blocked: stateCount.blocked,
-        sleeping: stateCount.sleeping,
-        unknown: stateCount.unknown,
-        zombie: stateCount.zombie,
-        list: topProcessesByCpu,
-      },
-    };
-  } catch (error) {
-    logger.error(`Failed to get process metrics: ${(error as Error).message}`);
-    throw error;
-  }
+    connectionCount: number;
+  };
 }
 
 // get the formatted cpu metrics
@@ -279,4 +129,5 @@ export const CollectorService = {
   getCpuUsagePerCoreMetrics,
   getDiskMetrics,
   getProcessMetrics,
+  getNetworkMetrics,
 };
